refactor(core): join core and core_web apps in a single phoenix app
