+++ apps/core/lib/core_web/views/error_view.ex
