+++ apps/core/test/unit/core_web/views/error_view_test.exs
