--- apps/core_web/.formatter.exs
@@ -1,4 +1,0 @@
-[
-  import_deps: [:phoenix],
-  inputs: ["*.{ex,exs}", "{config,lib,test}/**/*.{ex,exs}"]
-]
