+++ config/dev.exs
@@ -22,7 +22,7 @@ config :libcluster, debug: false
 # The watchers configuration can be used to run external
 # watchers to your application. For example, we use it
 # with esbuild to bundle .js and .css sources.
-config :core_web, CoreWeb.Endpoint,
+config :core, CoreWeb.Endpoint,
   # Binding to loopback ipv4 address prevents access from other machines.
   # Change to `ip: {0, 0, 0, 0}` to allow access from other machines.
   http: [ip: {127, 0, 0, 1}, port: 4000],
@@ -31,6 +31,13 @@ config :core_web, CoreWeb.Endpoint,
   debug_errors: true,
   secret_key_base: "YQWoCfQDbuv2yB8YdF+TXhDHI+9T3wJj0W1iT0A0yg+i7f0I0jyo3mtUvULET37c",
   watchers: []
+
+# Set a higher stacktrace during development. Avoid configuring such
+# in production as building large stacktraces may be expensive.
+config :phoenix, :stacktrace_depth, 20
+
+# Initialize plugs at runtime for faster development compilation
+config :phoenix, :plug_init_mode, :runtime
 
 # ## SSL Support
 #
