+++ config/prod.exs
@@ -28,7 +28,7 @@ config :logger,
 # manifest is generated by the `mix phx.digest` task,
 # which you should run after static files are built and
 # before starting your production server.
-config :core_web, CoreWeb.Endpoint, url: [host: "funless.dev", port: 80]
+config :core, CoreWeb.Endpoint, url: [host: "funless.dev", port: 80]
 # cache_static_manifest: "priv/static/cache_manifest.json"
 
 # ## SSL Support
@@ -36,7 +36,7 @@ config :core_web, CoreWeb.Endpoint, url: [host: "funless.dev", port: 80]
 # To get SSL working, you will need to add the `https` key
 # to the previous section and set your `:url` port to 443:
 #
-#     config :core_web, CoreWeb.Endpoint,
+#     config :core, CoreWeb.Endpoint,
 #       ...,
 #       url: [host: "example.com", port: 443],
 #       https: [
@@ -60,7 +60,7 @@ config :core_web, CoreWeb.Endpoint, url: [host: "funless.dev", port: 80]
 # We also recommend setting `force_ssl` in your endpoint, ensuring
 # no data is ever sent via http, always redirecting to https:
 #
-#     config :core_web, CoreWeb.Endpoint,
+#     config :core, CoreWeb.Endpoint,
 #       force_ssl: [hsts: true]
 #
 # Check `Plug.SSL` for all available options in `force_ssl`.
