+++ config/test.exs
@@ -32,7 +32,7 @@ config :worker, Worker.Domain.Ports.Runtime.Cleaner, adapter: Worker.Cleaner.Moc
 # --- Core Web Configs ---
 # We don't run a server during test. If one is required,
 # you can enable the server option below.
-config :core_web, CoreWeb.Endpoint,
+config :core, CoreWeb.Endpoint,
   http: [ip: {127, 0, 0, 1}, port: 4002],
   secret_key_base: "qUx9qUZ2rw23iFpaaPtFTwDhcULosXXK5l/wAv3o4MSHS0WYWNFC7D4v6m2e1pX7",
   server: false
